factor out client communication
