+++ dotnet-interactive-vscode/src/extension.ts
@@ -1,9 +1,11 @@
 import * as vscode from 'vscode';
 import { DotNetInteractiveNotebookProvider } from './notebookProvider';
+import { InteractiveClient } from './interactiveClient';
 import { registerLanguageProviders } from './languageProvider';
 
 export function activate(context: vscode.ExtensionContext) {
-    context.subscriptions.push(vscode.notebook.registerNotebookProvider('dotnet-interactive', new DotNetInteractiveNotebookProvider()));
+    let client = new InteractiveClient();
+    context.subscriptions.push(vscode.notebook.registerNotebookProvider('dotnet-interactive', new DotNetInteractiveNotebookProvider(client)));
     context.subscriptions.push(registerLanguageProviders());
 }
